simplify qr code stuff
